requirements.txt
pandas==0.24.2
loguru==0.5.1
opencv_python==4.1.2.30
numpy==1.18.1
scipy==1.4.1
tqdm==4.43.0
torch==1.4.0
matplotlib==3.2.1
torchvision==0.5.0
augmentor==0.2.8
ipython==7.16.1
Pillow==7.2.0
scikit_learn==0.23.1
